+++ Runtime/GameFeedBacks/Destroy.cs
@@ -1,16 +1,15 @@
-﻿using System.Threading.Tasks;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace Etienne.Feedback
 {
     [GameFeedback(204, 0, 0, "Destroy")]
     public class Destroy : GameFeedback
     {
-        public override async Task Execute(GameObject gameObject)
+        protected override void Execute(GameObject gameObject)
         {
             UnityEngine.GameObject.Destroy(gameObject);
-            await Task.CompletedTask;
         }
+
         public override string ToString()
         {
             return "Destroy";