+++ Runtime/GameFeedBacks/GameEvent.cs
@@ -1,3 +1,4 @@
+using System.Collections;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using UnityEngine;
@@ -10,13 +11,28 @@ namespace Etienne.Feedback
         public List<GameFeedback> Feedbacks => feedbacks;
         [SerializeReference] private List<GameFeedback> feedbacks;
 
-        public async Task Execute(GameObject gameObject)
+        public async Task ExecuteAsync(GameObject gameObject)
         {
+#if UNITY_WEBGL
+            Debug.LogWarning("WebGL doesn't support multi threaded behaviours, consider using ExecuteCoroutine in a Coroutine instead.");
+#endif
             foreach(GameFeedback feedback in Feedbacks)
             {
                 if(!Application.isPlaying || gameObject == null) return;
-                await feedback.Execute(gameObject);
+                await feedback.ExecuteAsync(gameObject);
             }
+        }
+
+        /// <summary>
+        /// Use With StartCoroutine()
+        /// </summary>
+        public IEnumerator ExecuteCoroutine(GameObject gameObject)
+        {
+            foreach(GameFeedback feedback in feedbacks)
+            {
+                yield return feedback.ExecuteCoroutine(gameObject);
+            }
+            yield break;
         }
     }
 }