+++ Runtime/GameFeedBacks/GameFeedback.cs
@@ -1,10 +1,22 @@
-using System.Threading.Tasks;
+using System.Collections;
 
 namespace Etienne.Feedback
 {
     [System.Serializable]
     public abstract class GameFeedback
     {
-        public abstract Task Execute(UnityEngine.GameObject gameObject);
+        protected abstract void Execute(UnityEngine.GameObject gameObject);
+
+        public virtual async System.Threading.Tasks.Task ExecuteAsync(UnityEngine.GameObject gameObject)
+        {
+            Execute(gameObject);
+            await System.Threading.Tasks.Task.CompletedTask;
+        }
+
+        public virtual IEnumerator ExecuteCoroutine(UnityEngine.GameObject gameObject)
+        {
+            Execute(gameObject);
+            yield break;
+        }
     }
 }